Build with QT_NO_NARROWING_CONVERSIONS_IN_CONNECT
Prevents floating-point-to-integral conversions in connects or mixing
up (u)ints.

+++ src/kcms/decoration/declarative-plugin/previewclient.cpp
@@ -426,8 +426,8 @@ void PreviewClient::requestToggleShade()
 
 SETTER(const QString &, setCaption, caption)
 SETTER(const QString &, setIconName, iconName)
-SETTER(int, setWidth, width)
-SETTER(int, setHeight, height)
+SETTER(qreal, setWidth, width)
+SETTER(qreal, setHeight, height)
 
 SETTER2(setActive, active)
 SETTER2(setCloseable, closeable)
